Segregate atomic elements the way WL does
Also, expand the AtomQ examples.
