+++ mathics/builtin/arithfns/basic.py
@@ -11,6 +11,7 @@ import sympy
 import mpmath
 
 from mathics.builtin.arithmetic import _MPMathFunction, create_infix
+from mathics.builtin.atomic.numbers import apply_N
 from mathics.builtin.base import (
     Builtin,
     BinaryOperator,
@@ -49,8 +50,6 @@ from mathics.core.systemsymbols import (
 from mathics.core.number import min_prec, dps
 
 from mathics.core.convert import from_sympy
-
-from mathics.builtin.numeric import apply_N
 
 from mathics.core.attributes import (
     flat,
