+++ mathics/builtin/atomic/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+Atomic Elements of Expresssions
+
+Expressions are ultimately built from a small number of distinct types of atomic elements.
+"""
