+++ mathics/builtin/drawing/plot.py
@@ -12,7 +12,13 @@ import itertools
 import palettable
 
 
-from mathics.core.expression import Expression
+from mathics.builtin.atomic.numbers import chop, apply_N
+from mathics.builtin.base import Builtin
+from mathics.builtin.graphics import Graphics
+from mathics.builtin.drawing.graphics3d import Graphics3D
+from mathics.builtin.options import options_to_rules
+from mathics.builtin.scoping import dynamic_scoping
+
 from mathics.core.atoms import (
     Real,
     MachineReal,
@@ -21,19 +27,12 @@ from mathics.core.atoms import (
     Integer0,
     from_python,
 )
+from mathics.core.attributes import hold_all, protected
+from mathics.core.expression import Expression
 from mathics.core.symbols import Symbol, SymbolList, SymbolN
 from mathics.core.systemsymbols import (
     SymbolRule,
 )
-
-from mathics.builtin.base import Builtin
-from mathics.builtin.graphics import Graphics
-from mathics.builtin.drawing.graphics3d import Graphics3D
-from mathics.builtin.numeric import chop, apply_N
-from mathics.builtin.options import options_to_rules
-from mathics.builtin.scoping import dynamic_scoping
-
-from mathics.core.attributes import hold_all, protected
 
 
 try:
