+++ mathics/builtin/files_io/filesystem.py
@@ -15,8 +15,8 @@ import time
 from mathics.builtin.base import Builtin, MessageException, Predefined
 from mathics.builtin.files_io.files import INITIAL_DIR  # noqa is used via global
 from mathics.builtin.files_io.files import DIRECTORY_STACK, MathicsOpen
-from mathics.builtin.numeric import Hash
-from mathics.builtin.strings import to_regex
+from mathics.builtin.atomic.numbers import Hash
+from mathics.builtin.atomic.strings import to_regex
 from mathics.core.atoms import Integer, Real, String, from_python
 from mathics.core.attributes import listable, locked, nothing, protected, read_protected
 from mathics.core.expression import Expression
