+++ mathics/builtin/lists.py
@@ -25,6 +25,8 @@ from mathics.algorithm.clusters import (
     optimize,
 )
 
+from mathics.builtin.atomic.numbers import apply_N
+
 from mathics.builtin.base import (
     Builtin,
     CountableInteger,
@@ -45,7 +47,6 @@ from mathics.builtin.exceptions import (
 from mathics.builtin.numbers.algebra import cancel
 from mathics.builtin.options import options_to_rules
 from mathics.builtin.scoping import dynamic_scoping
-from mathics.builtin.numeric import apply_N
 
 from mathics.core.atoms import (
     ByteArrayAtom,
